Make yr_scanner_define_boolean_variable receive an int instead of an int64_t. This keep the coherency with yr_compiler_define_boolean_variable and yr_rules_define_boolean_variable.

+++ libyara/include/yara/scanner.h
@@ -71,7 +71,7 @@ YR_API int yr_scanner_define_integer_variable(
 YR_API int yr_scanner_define_boolean_variable(
     YR_SCANNER* scanner,
     const char* identifier,
-    int64_t value);
+    int value);
 
 
 YR_API int yr_scanner_define_float_variable(
